+++ src/fractals/Julia.h
@@ -9,7 +9,7 @@ public:
 
     void render() override;
     void menu() override;
-    std::pair<float, float> coordsForZoom() const override;
+    std::pair<double, double> coordsForZoom() const override;
 
 private:
     float m_v1 = 0.311f;
